Fixed many build warnings.

+++ source/quake/keys.cpp
@@ -141,7 +141,7 @@ void Key_Console (int key)
 	if (key == K_PGUP || key==K_MWHEELUP)
 	{
 		con_backscroll += 2;
-		if (con_backscroll > con_totallines - (vid.height>>3) - 1)
+		if (con_backscroll > con_totallines - ((int)vid.height>>3) - 1)
 			con_backscroll = con_totallines - (vid.height>>3) - 1;
 		return;
 	}
